+++ weave/legacy/ops_domain/table.py
@@ -5,7 +5,7 @@ import json
 import logging
 import typing
 
-from weave import util, weave_internal
+from weave import weave_internal
 from weave import weave_types as types
 from weave.legacy.api import op, weave_class
 from weave.legacy import (
@@ -16,6 +16,7 @@ from weave.legacy import (
     wandb_util,
     engine_trace,
     errors, 
+    util, 
 )
 from weave.legacy import timestamp as weave_timestamp
 from weave.legacy.ops_domain import trace_tree, wbmedia
