+++ weave/legacy/panels_py/panel_autoboard.py
@@ -26,9 +26,10 @@ import os
 import typing
 
 import weave
-from weave import util, weave_internal
+from weave import weave_internal
 from weave.legacy.panels import panel_plot
 from weave.legacy.panels_py.generator_templates import template_registry
+from weave.legacy import util
 
 
 @weave.type()
