+++ weave/tests/fixture_fakewandb.py
@@ -11,12 +11,11 @@ from urllib import parse
 import wandb
 
 import weave
-from weave import util
 
 # Note: We're mocking out the whole io_service right now. This is too
 # high level and doesn't test the actual io implementation. We should
 # mock wandb_api instead probably.
-from weave.legacy import io_service, wandb_api, wandb_client_api
+from weave.legacy import io_service, util, wandb_api, wandb_client_api
 from weave.legacy.artifact_wandb import (
     WandbArtifact,
     WandbArtifactManifest,
